Add Amazon s3 support for archives service.
This commit adds support for listing remote archived WALs stored in
Amazon s3 for the archives service. A new boolean arg `--repo-s3` is
added to enable s3 support. This requires the `--config` arg so that we
can parse the AWS Access/Secret keys and bucket name which should
already be configured if using s3 support in pgBackRest. The
`--repo-path` arg should be set to the path prefix within the s3 bucket.

The following perl modules are required:

- Net::Amazon::S3
- Config::IniFiles

These can be easily installed on debian/ubuntu systems with:

```
$ apt-get install libnet-amazon-s3-perl libconfig-inifiles-perl
```

Example usage:

```
$ check_pgbackrest --stanza=demo --service=archives --format=human \
      --repo-s3 --repo-path /pgbackrest/archive \
      --config /etc/pgbackrest.conf
```